feat(breaking): Transactions to only include job name
Before, the title of a transaction would include both the job name and the workflow name.
After this change, we only take the job name and store the workflow path (unique unlike the name) as a tag.
This will allow listing jobs that belong to the same workflow.

We also include some extra metadata that allows right clicking and navigate to Github. Easier to investigate issues.