Fix ingestion from test fixture (#10)
In order to write tests, I had a generated test fixture that had all spans with the same uuid.

Unfortunately, a trace with duplicated span IDs will not be ingested by Sentry.

Fixing this does not fix the app since we ingest real workflow events. This has the learning value on how to write tests closer to reality.